+++ backend/esap/src/main/java/ru/javavlsu/kb/esap/service/PatientService.java
@@ -37,7 +37,7 @@ public class PatientService {
         return patientMapper.toPatientResponseDTOList(patients);
     }
 
-    public Long getPatientCountByClinic(Clinic clinic) {
+    public int getPatientCountByClinic(Clinic clinic) {
         return patientRepository.countPatientByClinic(clinic);
     }
